+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/definition/NodeDependency.scala
@@ -71,8 +71,8 @@ object Parameter {
 
   def apply(name: ParameterName, typ: TypingResult, validators: List[ParameterValidator]): Parameter =
     Parameter(
-      name,
-      typ,
+      name = name,
+      typ = typ,
       editors = Nil,
       validators = validators,
       defaultValue = None,
@@ -86,6 +86,7 @@ object Parameter {
       labelOpt = None,
       category = ParameterCategory.Standard,
       changesCanReloadParameters = false,
+      nonImportantForExecution = false,
     )
 
   def optional[T: TypeTag: NotNothing](name: ParameterName): Parameter =
@@ -95,8 +96,8 @@ object Parameter {
   // you should redefine scalaOptionParameter and javaOptionalParameter
   def optional(name: ParameterName, typ: TypingResult): Parameter =
     Parameter(
-      name,
-      typ,
+      name = name,
+      typ = typ,
       editors = Nil,
       validators = List.empty,
       defaultValue = None,
@@ -110,6 +111,7 @@ object Parameter {
       labelOpt = None,
       category = ParameterCategory.Standard,
       changesCanReloadParameters = false,
+      nonImportantForExecution = false,
     )
 
 }
@@ -142,7 +144,8 @@ case class Parameter(
     hintText: Option[String],
     labelOpt: Option[String],
     category: ParameterCategory,
-    changesCanReloadParameters: Boolean
+    changesCanReloadParameters: Boolean,
+    nonImportantForExecution: Boolean
 ) extends NodeDependency {
 
   def copy(
@@ -159,17 +162,17 @@ case class Parameter(
       javaOptionalParameter: Boolean,
   ): Parameter = {
     copy(
-      name,
-      typ,
-      editors,
-      validators,
-      defaultValue,
-      additionalVariables,
-      variablesToHide,
-      branchParam,
-      isLazyParameter,
-      scalaOptionParameter,
-      javaOptionalParameter,
+      name = name,
+      typ = typ,
+      editors = editors,
+      validators = validators,
+      defaultValue = defaultValue,
+      additionalVariables = additionalVariables,
+      variablesToHide = variablesToHide,
+      branchParam = branchParam,
+      isLazyParameter = isLazyParameter,
+      scalaOptionParameter = scalaOptionParameter,
+      javaOptionalParameter = javaOptionalParameter,
       hintText = None,
       labelOpt = None
     )
@@ -190,24 +193,26 @@ case class Parameter(
       hintText: Option[String] = this.hintText,
       labelOpt: Option[String] = this.labelOpt,
       category: ParameterCategory = this.category,
-      changesCanReloadParameters: Boolean = this.changesCanReloadParameters
+      changesCanReloadParameters: Boolean = this.changesCanReloadParameters,
+      nonImportantForExecution: Boolean = this.nonImportantForExecution,
   ): Parameter = {
     new Parameter(
-      name,
-      typ,
-      editors,
-      validators,
-      defaultValue,
-      additionalVariables,
-      variablesToHide,
-      branchParam,
-      isLazyParameter,
-      scalaOptionParameter,
-      javaOptionalParameter,
-      hintText,
-      labelOpt,
-      category,
-      changesCanReloadParameters
+      name = name,
+      typ = typ,
+      editors = editors,
+      validators = validators,
+      defaultValue = defaultValue,
+      additionalVariables = additionalVariables,
+      variablesToHide = variablesToHide,
+      branchParam = branchParam,
+      isLazyParameter = isLazyParameter,
+      scalaOptionParameter = scalaOptionParameter,
+      javaOptionalParameter = javaOptionalParameter,
+      hintText = hintText,
+      labelOpt = labelOpt,
+      category = category,
+      changesCanReloadParameters = changesCanReloadParameters,
+      nonImportantForExecution = nonImportantForExecution,
     )
   }
 
@@ -226,24 +231,26 @@ case class Parameter(
       hintText: Option[String],
       labelOpt: Option[String],
       category: ParameterCategory,
-      changesCanReloadParameters: Boolean
+      changesCanReloadParameters: Boolean,
+      nonImportantForExecution: Boolean,
   ): Parameter = {
     new Parameter(
-      name,
-      typ,
-      editors,
-      validators,
-      defaultValue,
-      additionalVariables,
-      variablesToHide,
-      branchParam,
-      isLazyParameter,
-      scalaOptionParameter,
-      javaOptionalParameter,
-      hintText,
-      labelOpt,
-      category,
-      changesCanReloadParameters
+      name = name,
+      typ = typ,
+      editors = editors,
+      validators = validators,
+      defaultValue = defaultValue,
+      additionalVariables = additionalVariables,
+      variablesToHide = variablesToHide,
+      branchParam = branchParam,
+      isLazyParameter = isLazyParameter,
+      scalaOptionParameter = scalaOptionParameter,
+      javaOptionalParameter = javaOptionalParameter,
+      hintText = hintText,
+      labelOpt = labelOpt,
+      category = category,
+      changesCanReloadParameters = changesCanReloadParameters,
+      nonImportantForExecution = nonImportantForExecution,
     )
   }
 
@@ -261,21 +268,22 @@ case class Parameter(
       javaOptionalParameter: Boolean,
   ): Parameter = {
     new Parameter(
-      name,
-      typ,
-      editors,
-      validators,
-      defaultValue,
-      additionalVariables,
-      variablesToHide,
-      branchParam,
-      isLazyParameter,
-      scalaOptionParameter,
-      javaOptionalParameter,
+      name = name,
+      typ = typ,
+      editors = editors,
+      validators = validators,
+      defaultValue = defaultValue,
+      additionalVariables = additionalVariables,
+      variablesToHide = variablesToHide,
+      branchParam = branchParam,
+      isLazyParameter = isLazyParameter,
+      scalaOptionParameter = scalaOptionParameter,
+      javaOptionalParameter = javaOptionalParameter,
       hintText = None,
       labelOpt = None,
       category = ParameterCategory.Standard,
       changesCanReloadParameters = false,
+      nonImportantForExecution = false,
     )
   }