+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/typeinformation/TypingResultAwareTypeInformationDetection.scala
@@ -7,7 +7,19 @@ import pl.touk.nussknacker.engine.api.{Context, ValueWithContext}
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.typed.typing._
 import pl.touk.nussknacker.engine.flink.api.TypedMultiset
-import pl.touk.nussknacker.engine.flink.api.typeinformation.{FlinkTypeInfoRegistrar, TypeInformationDetection}
+import pl.touk.nussknacker.engine.flink.api.typeinformation.{
+  CharsetTypeInformation,
+  CurrencyTypeInformation,
+  DurationTypeInformation,
+  FlinkTypeInfoRegistrar,
+  LocaleTypeInformation,
+  OffsetDateTimeTypeInformation,
+  PeriodTypeInformation,
+  TypeInformationDetection,
+  UUIDTypeInformation,
+  ZonedDateTimeTypeInformation,
+  ZoneIdTypeInformation
+}
 import pl.touk.nussknacker.engine.flink.typeinformation.ConcreteCaseClassTypeInfo
 import pl.touk.nussknacker.engine.process.typeinformation.internal.ContextTypeHelpers
 import pl.touk.nussknacker.engine.process.typeinformation.internal.typedobject.{
@@ -16,6 +28,9 @@ import pl.touk.nussknacker.engine.process.typeinformation.internal.typedobject.{
 }
 import pl.touk.nussknacker.engine.util.Implicits._
 
+import java.nio.charset.Charset
+import java.time.{Duration, OffsetDateTime, Period, ZonedDateTime, ZoneId}
+import java.util.{Currency, Locale, UUID}
 import scala.jdk.CollectionConverters._
 
 // TODO: handle avro types - see FlinkConfluentUtils
@@ -45,6 +60,15 @@ class TypingResultAwareTypeInformationDetection extends TypeInformationDetection
       case FlinkBelow119AdditionalTypeInfo(typeInfo) => typeInfo
       case TypedClass(klass, elementType :: Nil) if klass == classOf[java.util.List[_]] =>
         new ListTypeInfo[AnyRef](forType[AnyRef](elementType))
+      case TypedClass(klass, Nil) if klass == classOf[ZonedDateTime]                => ZonedDateTimeTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[OffsetDateTime]               => OffsetDateTimeTypeInformation
+      case TypedClass(klass, Nil) if classOf[ZoneId].isAssignableFrom(klass)        => ZoneIdTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[Duration]                     => DurationTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[Period]                       => PeriodTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[Charset]                      => CharsetTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[Currency]                     => CurrencyTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[Locale]                       => LocaleTypeInformation
+      case TypedClass(klass, Nil) if klass == classOf[UUID]                         => UUIDTypeInformation
       case TypedClass(klass, elementType :: Nil) if klass == classOf[Array[AnyRef]] =>
         // We have to use OBJECT_ARRAY even for numeric types, because ARRAY<INT> is represented as Integer[] which can't be handled by IntPrimitiveArraySerializer
         Types.OBJECT_ARRAY(forType[AnyRef](elementType))