+++ engine/flink/schemed-kafka-components-utils/src/test/scala/pl/touk/nussknacker/engine/schemedkafka/schema/AllTypesJsonSchemaWithoutDefaultValues.scala
@@ -62,9 +62,11 @@ object AllTypesJsonSchemaWithoutDefaultValues {
        |    },
        |    "unionField": {
        |      "anyOf": [
-       |        { "type": "null" },
        |        { "type": "string" },
-       |        { "type": "integer" }
+       |        { "type": "null" },
+       |        { "type": "integer" },
+       |        { "type": "boolean" },
+       |        { "type": "object" }
        |      ]
        |    },
        |    "recordField": {