+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/definition/component/DynamicComponentStaticDefinitionDeterminer.scala
@@ -13,10 +13,16 @@ import pl.touk.nussknacker.engine.api.context.transformation.{
 }
 import pl.touk.nussknacker.engine.api.definition.{OutputVariableNameDependency, Parameter}
 import pl.touk.nussknacker.engine.api.typed.typing.{TypingResult, Unknown}
-import pl.touk.nussknacker.engine.compile.nodecompilation.DynamicNodeValidator
+import pl.touk.nussknacker.engine.compile.nodecompilation.{
+  DynamicNodeValidator,
+  MultipleInputBranchesNodeInputValidationContext,
+  NodeInputValidationContext,
+  SingleInputNodeInputValidationContext
+}
 import pl.touk.nussknacker.engine.definition.component.DynamicComponentStaticDefinitionDeterminer.staticReturnType
 import pl.touk.nussknacker.engine.definition.component.dynamic.DynamicComponentDefinitionWithImplementation
 import pl.touk.nussknacker.engine.definition.component.parameter.StandardParameterEnrichment
+import pl.touk.nussknacker.engine.variables.GlobalVariablesPreparer
 
 // This class purpose is to provide initial set of parameters that will be presented after first usage of a component.
 // It is necessary to provide them, because:
@@ -24,7 +30,8 @@ import pl.touk.nussknacker.engine.definition.component.parameter.StandardParamet
 // - Sometimes user want to just use the component without filling parameters with own data - in this case we want to make sure
 //   that parameters will be available in the scenario, even with a default values
 class DynamicComponentStaticDefinitionDeterminer(
-    nodeValidator: DynamicNodeValidator
+    nodeValidator: DynamicNodeValidator,
+    globalVariablesPreparer: GlobalVariablesPreparer,
 ) extends LazyLogging {
 
   private def determineStaticDefinition(
@@ -40,20 +47,20 @@ class DynamicComponentStaticDefinitionDeterminer(
   private def determineInitialParameters(
       dynamic: DynamicComponentDefinitionWithImplementation
   )(implicit scenarioCompilationDependencies: ScenarioCompilationDependencies): List[Parameter] = {
-    def inferParameters(
-        transformer: DynamicComponent[_]
-    )(inputContext: transformer.InputContext) = {
+    def inferParameters(transformer: DynamicComponent[_], inputContext: NodeInputValidationContext) = {
       // We assume that this information is not important for determining initial parameters of dynamic nodes, so we pass fake values
       implicit val nodeId: NodeId = NodeId("fakeNodeId")
       nodeValidator
         .validateNode(
-          transformer,
-          Nil,
-          Nil,
-          if (dynamic.component.nodeDependencies.contains(OutputVariableNameDependency)) Some("fakeOutputVariable")
-          else None,
-          dynamic.parametersConfig
-        )(inputContext)
+          component = transformer,
+          parametersFromNode = Nil,
+          branchParametersFromNode = Nil,
+          outputVariable =
+            if (dynamic.component.nodeDependencies.contains(OutputVariableNameDependency)) Some("fakeOutputVariable")
+            else None,
+          parametersConfig = dynamic.parametersConfig,
+          nodeInputValidationContext = inputContext
+        )
         .map(_.parameters)
         .valueOr { err =>
           logger.warn(
@@ -65,13 +72,19 @@ class DynamicComponentStaticDefinitionDeterminer(
         }
     }
 
+    val globalVariablesOnlyValidationContext =
+      globalVariablesPreparer.prepareValidationContextWithGlobalVariablesOnly(Set.empty)
+
     dynamic.component match {
       case withStatic: WithStaticParameters =>
         StandardParameterEnrichment.enrichParameterDefinitions(withStatic.staticParameters, dynamic.parametersConfig)
       case single: SingleInputDynamicComponent[_] =>
-        inferParameters(single)(ValidationContext())
+        inferParameters(single, SingleInputNodeInputValidationContext(globalVariablesOnlyValidationContext))
       case join: JoinDynamicComponent[_] =>
-        inferParameters(join)(Map.empty)
+        inferParameters(
+          join,
+          MultipleInputBranchesNodeInputValidationContext(Map.empty, globalVariablesOnlyValidationContext)
+        )
     }
   }
 
@@ -87,7 +100,10 @@ object DynamicComponentStaticDefinitionDeterminer {
   ): Map[ComponentId, ComponentStaticDefinition] = {
     val nodeValidator = DynamicNodeValidator(modelDataForType)
     val toStaticComponentDefinitionTransformer =
-      new DynamicComponentStaticDefinitionDeterminer(nodeValidator)
+      new DynamicComponentStaticDefinitionDeterminer(
+        nodeValidator,
+        GlobalVariablesPreparer(modelDataForType.modelDefinition.expressionConfig)
+      )
 
     // We have to wrap this block with model's class loader because it invokes node compilation under the hood
     modelDataForType.withModelClassloaderAsContextClassLoader {